Set precision of floating points parameters to double
floats are just not cutting it when working with data with large values,
e.g. in geographic frames of reference.

+++ libs/control/projectactions.cpp
@@ -160,7 +160,7 @@ bool estimateLocalCoordinateSystem(Project* p) {
 
     mean /= count;
 
-    p->setLocalCoordinateFrame(getLocalFrameOnSphere(mean, minDist));
+    p->setLocalCoordinateFrame(getLocalFrameOnSphere(mean, minDist).cast<double>());
 
     return true;
 }
